chore: rename `format` to `style`
