+++ src/config.ts
@@ -19,7 +19,7 @@ const defaultConfig = {
   capitalize: true,
   group: true,
   tag: 'v%s',
-  format: 'markdown',
+  style: 'markdown',
 } satisfies ChangelogOptions
 
 export async function resolveConfig(options: ChangelogOptions) {
