+++ src/index.ts
@@ -2,6 +2,6 @@ export * from './config'
 export * from './generate'
 export * from './git'
 export * from './github'
-export * from './format/markdown'
 export * from './parse'
+export * from './style/markdown'
 export * from './types'
